+++ init.go
@@ -3,6 +3,8 @@ package main
 import (
 	"encoding/json"
 	"os"
+	"path/filepath"
+	"plugin"
 	"strconv"
 	"strings"
 	"time"
@@ -11,6 +13,7 @@ import (
 	"github.com/Showmax/go-fqdn"
 	"github.com/sirupsen/logrus"
 
+	"github.com/AVENTER-UG/mesos-m3s/redis"
 	cfg "github.com/AVENTER-UG/mesos-m3s/types"
 )
 
@@ -166,6 +169,48 @@ func init() {
 		config.K3SEnableTaint = true
 	}
 
+	// Enable plugins
+	if strings.Compare(util.Getenv("M3S_PLUGINS", "false"), "false") == 0 {
+		config.PluginsEnable = false
+	} else {
+		config.PluginsEnable = true
+	}
+}
+
+func loadPlugins(r *redis.Redis) {
+	if config.PluginsEnable {
+		config.Plugins = map[string]*plugin.Plugin{}
+
+		plugins, err := filepath.Glob("plugins/*.so")
+		if err != nil {
+			logrus.WithField("func", "main.loadPlugins").Info("No Plugins found")
+			return
+		}
+
+		for _, filename := range plugins {
+			p, err := plugin.Open(filename)
+			if err != nil {
+				logrus.WithField("func", "main.initPlugins").Error("Error during loading plugin: ", err.Error())
+				continue
+			}
+
+			symbol, err := p.Lookup("Init")
+			if err != nil {
+				logrus.WithField("func", "main.initPlugins").Error("Error lookup init plugin: ", err.Error())
+				continue
+			}
+
+			initPluginFunc, ok := symbol.(func(*redis.Redis) string)
+
+			if !ok {
+				logrus.WithField("func", "main.initPlugins").Error("Error plugin does not have init function")
+				continue
+			}
+
+			name := initPluginFunc(r)
+			config.Plugins[name] = p
+		}
+	}
 }
 
 func stringToBool(par string) bool {