+++ shell.nix
@@ -1,9 +1,18 @@
-{ pkgs ? import <nixpkgs> { } }:
+with import <nixpkgs> {};
 
-with pkgs;
+stdenv.mkDerivation {
+name = "go-env";
 
-mkShell {
-  buildInputs = [
-    go
-  ];
+buildInputs = [
+  go
+];
+
+SOURCE_DATE_EPOCH = 315532800;
+PROJDIR = "${toString ./.}";
+
+shellHook = ''
+    export LD_LIBRARY_PATH="${pkgs.stdenv.cc.cc.lib}/lib"
+    export PATH=/tmp/bin:$PATH
+    mkdir /tmp/bin
+    '';
 }